Fix(community): FIxed community locust tests #85
